% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data(): Clean the dataset}
\usage{
eq_clean_data(dataset)
}
\arguments{
\item{dataset}{Dataset to be kept tidy}
}
\value{
Simplified dataset
}
\description{
This function does three chores including conversion of datatypes, making date column, and drop unneeded columns.
More specifically, latitude and longitude columns are converted to numeric type.
Date column is made by combining year, month, and day column, and converted to datetime type.
Finally, only columns id, year, date, country, location_name, longitude, latitude, deaths, total_deaths, eq_primary are preserved.
}
\examples{
\dontrun{
eq_load_data('earthquakes.tsv.gz') \%>\%
eq_clean_data()
}
}
