% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_leaftlets.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label(): Create an annotated text to be used in leaflet pop-up}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Dataframe from which the function extracts necessary infos}
}
\value{
Created popup_text
}
\description{
This function creates a html tag including three lines of data, which pops up when the user clicks on a location.
The information includes location, magnitude, and total deaths.
Using <br> and <b> tags, the text is shown in three lines with subjects emphasized (made bold-faced).
}
\examples{
\dontrun{
dataset$popup_text <- eq_load_data('earthquake.tsv.gz') \%>\%
eq_create_label()
}
}
