% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_load_data}
\alias{eq_load_data}
\title{eq_load_data(): Read raw NOAA datafile with readr::read_delim() function.}
\usage{
eq_load_data(filename)
}
\arguments{
\item{filename}{String-type file name provided and results in dataframe}
}
\value{
This function returns a dataframe object
}
\description{
This function accepts a filename and reads NOAA earthquake data file, thereby creating dataframe.
The function is capable of cheking if the file exists and prints error msg when it doesn't.
}
\examples{
\dontrun{
eq_load_data('earthquakes.tsv.gz')
}
}
