% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean(): Clean out location name
This function separates out country name from the location name including colon(:).
To achieve the objective, it employs regular expressions and finally converts the chracters to the title form.}
\usage{
eq_location_clean(data_col)
}
\arguments{
\item{data_col}{location_name column to be simplified}
}
\value{
Simplified column
}
\description{
eq_location_clean(): Clean out location name
This function separates out country name from the location name including colon(:).
To achieve the objective, it employs regular expressions and finally converts the chracters to the title form.
}
\examples{
\dontrun{
eq_load_data('earthquakes.tsv.gz') \%>\%
eq_location_clean(location_name)
}
}
