% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_leaftlets.R
\name{eq_map}
\alias{eq_map}
\title{eq_map(): Build a leaflet map indicating earthquake locations}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{Dataframe from which the map is drawn}

\item{annot_col}{Column to be used for popup info}
}
\value{
Leaflet map object to be drawn
}
\description{
This function builds an interactive leaflet map indicating locations of earthquakes indicated by points on the map.
It can also pull out information about the quake using user-provided columns.
Examples include data and popup_text columns, latter of which shows three-lined summary info.
}
\examples{
\dontrun{
dataset$popup_text <- eq_load_data('earthquake.tsv.gz') \%>\%
eq_map(date)

dataset <- eq_create_label(dataset)
eq_map(dataset,'popup_text')
}
}
