% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_lable(): geom function to write label of location infos}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{a set of aesthetic mappings}

\item{data}{data to be plotted}

\item{stat}{stat object (No custom version used here)}

\item{position}{position object (No custom version used here)}

\item{show.legend}{inclusion of the legend}

\item{na.rm}{treatment of missing values}

\item{inherit.aes}{inheriting aeses from default geom}

\item{...}{additional parameters}
}
\value{
None
}
\description{
geom_timeline_lable(): geom function to write label of location infos
}
\examples{
\dontrun{
x_min <- as.Date('2003-01-01',origin='1970-01-01')
x_max <- as.Date('2017-01-01',origin='1970-01-01')
eq_load_data('earthquakes.tsv.gz') \%>\%
eq_clean_data() \%>\%
filter(year >= 2000) \%>\%
filter(country \%in\% c("USA","MEXICO")) \%>\%
ggplot2::ggplot(ggplot2::aes(x=date,y=country,colour=deaths,size=eq_primary)) +
geom_timeline(aes(xmin=x_min, xmax=x_max)) +
geom_timeline_label(aes(xmin=x_min,xmax=x_max,label=location_name,n_max=5))
}
}
