% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline: Geom for drawing earthquakes on the timeline}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimeline
}
\description{
This Geom take earthquake dataset and can show magnitudes and deaths by circles on a line.
Only x aes, representing date of earthquake is mandatory.
However, size and colour aeses can be used to represent magnitude of and deaths from the earthquakes, respectively.
}
\details{
To make the customized range of minimum and maximum range of x axis, xmin and xmax aeses can be provided (datetime).
When not given, the minimum and maximum value of date object are calculated by default.
}
\keyword{datasets}
