% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel: Geom for adding text labels of earthquake locations on the timeline}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimelineLabel
}
\description{
This Geom is the optional addon for GeomTimeline Geom element.
This aims to give extra information of locations by adding text labels on top of the pointsGrob element.
The text is 45 degree slanted and takes three mandatory aeses, x, label, and n_max.
The aes x takes datetime type parameter, label takes the column which would be written, finally n_max is the number of allowed earthquake magnitudes.
In that way, this Geom takes top n_max earthquakes.
}
\details{
To make the same range of xmin and xmax, same logics are applied inside the draw_group function.
}
\keyword{datasets}
