% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{helper_neg_dates}
\alias{helper_neg_dates}
\title{helper_neg_dates(): Converts BC-dated character to datetime object}
\usage{
helper_neg_dates(date)
}
\arguments{
\item{date}{A BC date to be converted}
}
\value{
Converted BC datetime object
}
\description{
This function mitigates the shortfall of lubridate::ymd() and as.Date() as they fail to convert BC dates.
The function accepts a variale of BC-dated character and mirror it with respect to origin of AD0.
Subraction of the date from that origin and add 1 day results in expected BC date.
}
\examples{
\dontrun{
helper_neg_dates('-2145-01-01')}
}
