% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{theme_eq_custom}
\alias{theme_eq_custom}
\title{theme_eq_custom(): Setting customized theme for simplifying unneeded elements}
\usage{
theme_eq_custom(...)
}
\arguments{
\item{...}{No parameters}
}
\value{
None but adjusting the apperance of ggplot object
}
\description{
This function specifies several graphical elements in a way in coordance with the guideline.
A few adjustments include placing the legend in the bottom, removing x and y axes grids, and changes colors to gray.
}
\examples{
\dontrun{
x_min <- as.Date('2003-01-01',origin='1970-01-01')
x_max <- as.Date('2017-01-01',origin='1970-01-01')
eq_load_data('earthquakes.tsv.gz') \%>\%
eq_clean_data() \%>\%
filter(year >= 2000) \%>\%
filter(country \%in\% c("USA","MEXICO")) \%>\%
ggplot2::ggplot(ggplot2::aes(x=date,y=country,colour=deaths,size=eq_primary)) +
geom_timeline(aes(xmin=x_min, xmax=x_max)) +
geom_timeline_label(aes(xmin=x_min,xmax=x_max,label=location_name,n_max=5)) +
ggplot2::scale_x_date(name = "DATE") +
ggplot2::scale_size_continuous(name='Richter scale value') +
ggplot2::scale_color_continuous(name='# of deaths') +
theme_eq_custom()
}
}
